Stop depending on speccy and redoc
When you will update API documents, they will be installed temporarily.

We have documentation in Japanese, English, Simplified Chinese, and
Traditional Chinese.

We also have documentation for Splatoon 1 and Splatoon 2 in each of
these languages.

Due to the temporary installation for each of those 4×2 iterations,
updating the documentation has taken a lot of time.

However, it is expected that this commit will simplify dependencies.

Historically, dependency on redoc has resulted in unnecessary security
warnings. Redoc is only used on the developer's device. It won't raise a
real security issue on production server. However, GitHub's security
scanner doesn't know about this, so it can't be helped.

Originally, there is a --dev switch to avoid this situation, but the
lockfile implementation is broken.