Stop using LINE Seed EN and use JP
"LINE Seed EN" kerning process is not so beautiful and "LINE Seed JP"
gives better results, so change to use JP in English.

Unfortunately, Noto Sans is still in use for Latin languages other than
English, since LINE Seed does not have enough characters to write Latin
languages other than English, whether JP or EN.

LINE Seed JP contains Cyrillic characters, but which are obviously not
appropriate for Russian users, and therefore cannot be used for Russian
as well. (It contains full-width characters, which is ugly. The font
weight is also obviously different from that of LINE Seed's. This is
probably due to a mistake made by the LINE Corporation to include this
data.)